initial work on making PageHeader use asNavLink
