+++ components/page-header/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.1",
   "dependencies": {
     "@govuk-react/constants": "^0.2.7",
+    "as-nav-link": "^0.1.1",
     "govuk-colours": "^1.0.3"
   },
   "devDependencies": {
