+++ components/page-header/src/atoms/logo-anchor.js
@@ -1,0 +1,16 @@
+import styled from 'react-emotion';
+
+const PageHeaderAnchor = styled(({
+  as: T = 'a',
+}))({
+  fontSize: '30px',
+  lineHeight: '30px',
+  fontWeight: '700',
+});
+
+export default PageHeaderAnchor;
+
+const StyledLogoAnchor = styled(Anchor)({
+
+});
+return StyledLogoAnchor;
